♻️ LowerDashbordをPropsをとるように実装 (#49)
